House / guilds redesign

+++ src/game/server/core/mmo_controller.h
@@ -94,7 +94,7 @@ public:
 	//
 	void LoadLogicWorld() const;
 	void SaveAccount(CPlayer *pPlayer, int Table) const;
-	void ShowLoadingProgress(const char* pLoading, int Size) const;
+	void ShowLoadingProgress(const char* pLoading, size_t Size) const;
 	void ShowTopList(int ClientID, ToplistType Type, int Rows, class VoteWrapper* pWrapper = nullptr) const;
 };
